+++ src/contexts/AuthContext.tsx
@@ -74,6 +74,7 @@ export function AuthProvider({ children }: AuthProviderProps) {
     );
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export function useAuth(): AuthContextType {
     const context = useContext(AuthContext);
     if (!context) {
